+++ src/quick-lint-js/diag/diagnostic-metadata-generated.h
@@ -290,6 +290,8 @@ namespace quick_lint_js {
   QLJS_DIAG_TYPE_NAME(Diag_TypeScript_Enum_Is_Not_Allowed_In_JavaScript) \
   QLJS_DIAG_TYPE_NAME(Diag_TypeScript_Enum_Member_Name_Cannot_Be_Number) \
   QLJS_DIAG_TYPE_NAME(Diag_TypeScript_Enum_Value_Must_Be_Constant) \
+  QLJS_DIAG_TYPE_NAME(Diag_TypeScript_Export_As_Namespace_Is_Not_Allowed_In_Namespace_Or_Module) \
+  QLJS_DIAG_TYPE_NAME(Diag_TypeScript_Export_As_Namespace_Is_Only_Allowed_In_TypeScript_Definition_File) \
   QLJS_DIAG_TYPE_NAME(Diag_TypeScript_Export_Equal_Not_Allowed_In_JavaScript) \
   QLJS_DIAG_TYPE_NAME(Diag_TypeScript_Import_Type_Missing_Export_Name) \
   QLJS_DIAG_TYPE_NAME(Diag_TypeScript_Implements_Must_Be_After_Extends) \
@@ -444,7 +446,7 @@ namespace quick_lint_js {
   /* END */
 // clang-format on
 
-inline constexpr int Diag_Type_Count = 430;
+inline constexpr int Diag_Type_Count = 432;
 
 extern const Diagnostic_Info all_diagnostic_infos[Diag_Type_Count];
 }