+++ docs/CHANGELOG.md
@@ -6,6 +6,13 @@ based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 quick-lint-js' version numbers are arbitrary. quick-lint-js does *not* adhere to
 Semantic Versioning.
 
+## Unreleased
+
+### Added
+
+* TypeScript support (still experimental):
+  * `export as namespace` statements are now parsed.
+
 ## 2.18.0 (2023-11-03)
 
 [Downloads](https://c.quick-lint-js.com/releases/2.18.0/)